Docker/app/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
click==8.0.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.28
protobuf==3.19.4
pydantic==1.9.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6